Marvel API DTO fusion to simplify
