+++ TechTest.xcodeproj/xcuserdata/xavier.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -10,5 +10,23 @@
 			<integer>5</integer>
 		</dict>
 	</dict>
+	<key>SuppressBuildableAutocreation</key>
+	<dict>
+		<key>254C217E27F85752005F900F</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>254C219427F85753005F900F</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>254C219E27F85753005F900F</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+	</dict>
 </dict>
 </plist>
