+++ TechTest/API Client/ApiClient.swift
@@ -4,4 +4,5 @@ import Combine
 protocol APIClient {
     func getCharacters() -> AnyPublisher<[Marvel.Character], Error>
     func getCharacters(nameStartsWith: String?) -> AnyPublisher<[Marvel.Character], Error>
+//    func getCharacters(onCompletion: @escaping (Result<[Marvel.Character], Error>) -> Void)
 }
